+++ embed_grader/serapis/utils/visualizer_manager.py
@@ -52,8 +52,8 @@ class VisualizerManager(object):
     def _get_visualizer(self, field_name, raw_content, visualizer_id):
         params = (raw_content, visualizer_id)
         if field_name.endswith('.stm32.waveform'):
-            from serapis.utils.visualizer.stm32_waveform_visualizer import STM32WaveformVisualizer
+            from serapis.utils.visualizers.stm32_waveform_visualizer import STM32WaveformVisualizer
             return STM32WaveformVisualizer(*params)
         else:
-            from serapis.utils.visualizer.plain_text_visualizer import PlainTextVisualizer
+            from serapis.utils.visualizers.plain_text_visualizer import PlainTextVisualizer
             return PlainTextVisualizer(*params)
